Fix #6209 - introduce standalone dir crate
Done so far:

* created the dir crate in util
* moved code from `ethstore/src/dir/paths.rs` to dir crate
* rename `dir` module in ethstore to `accounts_dir` to distinguish it
  from the dir crate

Is there a need to mutualize some code in the newly created
`util/dir/src/lib.rs` ? If so, I would need some guidance as of which
functions are to be mutualized.
