+++ ethstore/Cargo.toml
@@ -19,6 +19,7 @@ itertools = "0.5"
 parking_lot = "0.4"
 ethcrypto = { path = "../ethcrypto" }
 ethcore-bigint = { path = "../util/bigint" }
+dir = { path = "../util/dir" }
 smallvec = "0.4"
 parity-wordlist = "1.0"
 tempdir = "0.3"
