+++ ethstore/cli/Cargo.toml
@@ -9,6 +9,7 @@ serde = "1.0"
 serde_derive = "1.0"
 docopt = "0.8"
 ethstore = { path = "../" }
+dir = { path = '../../util/dir' }
 panic_hook = { path = "../../panic_hook" }
 
 [[bin]]
