+++ rpc/src/v1/tests/mocked/parity_accounts.rs
@@ -18,7 +18,7 @@ use std::sync::Arc;
 
 use ethcore::account_provider::{AccountProvider, AccountProviderSettings};
 use ethstore::EthStore;
-use ethstore::dir::RootDiskDirectory;
+use ethstore::accounts_dir::RootDiskDirectory;
 use devtools::RandomTempPath;
 
 use jsonrpc_core::IoHandler;
