+++ util/dir/src/lib.rs
@@ -14,7 +14,12 @@
 // You should have received a copy of the GNU General Public License
 // along with Parity.  If not, see <http://www.gnu.org/licenses/>.
 
-use std::fs;
+extern crate app_dirs;
+extern crate ethcore_bigint as bigint;
+extern crate journaldb;
+
+pub mod helpers;
+use std::{env, fs};
 use std::path::{PathBuf, Path};
 use bigint::hash::{H64, H256};
 use journaldb::Algorithm;
@@ -232,6 +237,82 @@ pub fn default_local_path() -> String {
 pub fn default_hypervisor_path() -> String {
 	let app_info = AppInfo { name: PRODUCT_HYPERVISOR, author: AUTHOR };
 	get_app_root(AppDataType::UserData, &app_info).map(|p| p.to_string_lossy().into_owned()).unwrap_or_else(|_| "$HOME/.parity-hypervisor".to_owned())
+}
+
+fn home() -> PathBuf {
+	env::home_dir().expect("Failed to get home dir")
+}
+
+/// Geth path
+pub fn geth(testnet: bool) -> PathBuf {
+	let mut base = geth_base();
+	if testnet {
+		base.push("testnet");
+	}
+	base.push("keystore");
+	base
+}
+
+/// Parity path for specific chain
+pub fn parity(chain: &str) -> PathBuf {
+	let mut base = parity_base();
+	base.push(chain);
+	base
+}
+
+#[cfg(target_os = "macos")]
+fn parity_base() -> PathBuf {
+	let mut home = home();
+	home.push("Library");
+	home.push("Application Support");
+	home.push("io.parity.ethereum");
+	home.push("keys");
+	home
+}
+
+#[cfg(windows)]
+fn parity_base() -> PathBuf {
+	let mut home = home();
+	home.push("AppData");
+	home.push("Roaming");
+	home.push("Parity");
+	home.push("Ethereum");
+	home.push("keys");
+	home
+}
+
+#[cfg(not(any(target_os = "macos", windows)))]
+fn parity_base() -> PathBuf {
+	let mut home = home();
+	home.push(".local");
+	home.push("share");
+	home.push("io.parity.ethereum");
+	home.push("keys");
+	home
+}
+
+#[cfg(target_os = "macos")]
+fn geth_base() -> PathBuf {
+	let mut home = home();
+	home.push("Library");
+	home.push("Ethereum");
+	home
+}
+
+#[cfg(windows)]
+fn geth_base() -> PathBuf {
+	let mut home = home();
+	home.push("AppData");
+	home.push("Roaming");
+	home.push("Ethereum");
+	home
+}
+
+#[cfg(not(any(target_os = "macos", windows)))]
+fn geth_base() -> PathBuf {
+	let mut home = home();
+	home.push(".ethereum");
+	home
 }
 
 #[cfg(test)]
